+++ SwrveConversationSDK/src/main/java/com/swrve/sdk/ISwrveConversationsSDK.java
@@ -6,5 +6,6 @@ import java.util.Map;
  * Used internally to interface the SwrveSDK (native and Unity) with conversations
  */
 interface ISwrveConversationsSDK {
+    int CONVERSATION_VERSION = 3;
     void queueConversationEvent(String viewEvent, String eventName, String page, String conversationId, Map<String, String> payload);
 }
