+++ SwrveSDK/src/main/java/com/swrve/sdk/SwrveBase.java
@@ -59,7 +59,7 @@ import java.util.concurrent.TimeUnit;
 /**
  * Main base class implementation of the Swrve SDK.
  */
-public abstract class SwrveBase<T, C extends SwrveConfigBase> extends SwrveImp<T, C> implements ISwrveBase<T, C>, ISwrveCommon {
+public abstract class SwrveBase<T, C extends SwrveConfigBase> extends SwrveImp<T, C> implements ISwrveBase<T, C>, ISwrveCommon, ISwrveConversationsSDK {
 
     protected SwrveBase(Context context, int appId, String apiKey, C config) {
         super(context, appId, apiKey, config);
@@ -579,7 +579,7 @@ public abstract class SwrveBase<T, C extends SwrveConfigBase> extends SwrveImp<T
                 deviceInfo.put(SWRVE_DEVICE_DPI, metrics.densityDpi);
                 deviceInfo.put(SWRVE_ANDROID_DEVICE_XDPI, xdpi);
                 deviceInfo.put(SWRVE_ANDROID_DEVICE_YDPI, ydpi);
-                deviceInfo.put(SWRVE_CONVERSATION_VERSION, CONVERSATION_VERSION);
+                deviceInfo.put(SWRVE_CONVERSATION_VERSION, ISwrveConversationsSDK.CONVERSATION_VERSION);
                 // Carrier info
                 if (!SwrveHelper.isNullOrEmpty(simOperatorName)) {
                     deviceInfo.put(SWRVE_SIM_OPERATOR_NAME, simOperatorName);
@@ -644,7 +644,7 @@ public abstract class SwrveBase<T, C extends SwrveConfigBase> extends SwrveImp<T
                 if (config.isTalkEnabled()) {
                     // Talk only params
                     params.put("version", String.valueOf(CAMPAIGN_ENDPOINT_VERSION));
-                    params.put("conversation_version", String.valueOf(CONVERSATION_VERSION));
+                    params.put("conversation_version", String.valueOf(ISwrveConversationsSDK.CONVERSATION_VERSION));
                     params.put("language", language);
                     params.put("app_store", config.getAppStore());
 
@@ -1632,11 +1632,6 @@ public abstract class SwrveBase<T, C extends SwrveConfigBase> extends SwrveImp<T
      */
 
     @Override
-    public ISwrveConversationsSDK getConversationSDK() {
-        return SwrveSDK.conversationInstance;
-    }
-
-    @Override
     public int getAppId() {
         return appId;
     }
@@ -1710,5 +1705,29 @@ public abstract class SwrveBase<T, C extends SwrveConfigBase> extends SwrveImp<T
 
     /***
      * eo ISwrveCommon
+     */
+
+    /***
+     * Implementation of ISwrveConversationsSDK methods
+     */
+
+    @Override
+    public void queueConversationEvent(String viewEvent, String eventName, String page, String conversationId, Map<String, String> payload) {
+        if (payload == null) {
+            payload = new HashMap<String, String>();
+        }
+        payload.put("event", eventName);
+        payload.put("conversation", conversationId);
+        payload.put("page", page);
+
+        SwrveLogger.d(LOG_TAG, "Sending view conversation event: " + viewEvent);
+
+        Map<String, Object> parameters = new HashMap<String, Object>();
+        parameters.put("name", viewEvent);
+        queueEvent("event", parameters, payload);
+    }
+
+    /***
+     * eo ISwrveConversationsSDK
      */
 }
